Move order related classes to separate folder
