+++ src/models/order.entity.ts
@@ -9,18 +9,19 @@ import {
 }                       from 'sequelize-typescript';
 import {
 	Field,
-	InterfaceType,
 	ObjectType
 }                       from '@nestjs/graphql';
 import { ApiProperty }  from '@nestjs/swagger';
 import { ORDER }        from '@config/json';
 import {
-	ActionStatus,
-	DestinationType,
 	LoadingType,
 	OrderStage,
 	OrderStatus
 }                       from '@common/enums';
+import {
+	OrderExecutionState,
+	OrderFilter
+}                       from '@common/classes/order';
 import { UuidScalar }   from '@common/scalars';
 import {
 	Reference,
@@ -36,8 +37,6 @@ import {
 	IntArrayColumn,
 	IntColumn,
 	IOrder,
-	IOrderExecutionState,
-	IOrderFilter,
 	JsonbColumn,
 	StringArrayColumn,
 	StringColumn,
@@ -69,92 +68,6 @@ const scopeOptions: FindOptions = {
 		}
 	]
 };
-
-@InterfaceType()
-export class OrderFilter
-	implements IOrderFilter {
-	/**
-	 * Minimal weight limit for order.
-	 * */
-	weightMin?: number;
-	/**
-	 * Maximal weight limit for order.
-	 * */
-	weightMax?: number;
-	/**
-	 * Minimal volume limit for order.
-	 * */
-	volumeMin?: number;
-	/**
-	 * Maximal volume limit for order.
-	 * */
-	volumeMax?: number;
-	/**
-	 * Minimal length limit for order.
-	 * */
-	lengthMin?: number;
-	/**
-	 * Maximal length limit for order.
-	 * */
-	lengthMax?: number;
-	/**
-	 * Minimal width limit for order.
-	 * */
-	widthMin?: number;
-	/**
-	 * Maximal width limit for order.
-	 * */
-	widthMax?: number;
-	/**
-	 * Minimal width limit for order.
-	 * */
-	heightMin?: number;
-	/**
-	 * Maximal height limit for order.
-	 * */
-	heightMax?: number;
-	/**
-	 * Types of transport for order.
-	 * */
-	types?: string[];
-	/**
-	 * Number of pallets of transport for order.
-	 * */
-	pallets?: number;
-	/**
-	 * For dedicated transports only.
-	 * */
-	isDedicated?: boolean;
-	/**
-	 * Capability for taking extra payload.
-	 * */
-	payloadExtra?: boolean;
-	/**
-	 * Payload filter.
-	 * */
-	payload?: string;
-	payloadType?: string;
-	loadingTypes?: LoadingType[];
-	status?: OrderStatus;
-	riskClass?: string;
-	paymentTypes?: string[];
-	dedicated?: string;
-	directions?: string[];
-	hasDriver?: boolean;
-	fromDate?: Date | string;
-	toDate?: Date | string;
-}
-
-@InterfaceType()
-export class OrderExecutionState
-	implements IOrderExecutionState {
-	type?: DestinationType;
-
-	actionStatus?: ActionStatus;
-	loaded?: boolean;
-	unloaded?: boolean;
-	uploaded?: boolean;
-}
 
 /**
  * Order model for cargo company
